+++ src/primitives.rs
@@ -44,20 +44,24 @@ macro_rules! impl_traits {
 
             #[inline]
             fn eq_abs(&self, other: &Self, max_diff: &Self::Epsilon) -> bool {
-                self.abs_diff(other).le(max_diff)
+                // the PartialEq check covers equality of infinities
+                self == other || self.abs_diff(other).le(max_diff)
             }
 
             #[inline]
             fn eq_rel(&self, other: &Self, max_diff: &Self::Epsilon) -> bool {
-                let largest = $float::abs(*self).max($float::abs(*other));
-                let epsilon = largest * max_diff;
-                self.abs_diff(other) <= epsilon
+                // the PartialEq check covers equality of infinities
+                self == other || {
+                    let largest = $float::abs(*self).max($float::abs(*other));
+                    let epsilon = largest * max_diff;
+                    self.abs_diff(other) <= epsilon
+                }
             }
 
             #[inline]
             fn eq_ulps(&self, other: &Self, max_diff: &Self::UlpsEpsilon) -> bool {
                 if self.is_nan() || other.is_nan() {
-                    false
+                    false // NaNs are never equal
                 } else if self.is_sign_positive() != other.is_sign_positive() {
                     self == other // account for zero == negative zero
                 } else {
@@ -158,8 +162,7 @@ impl_traits!(f64, u64);
 // tests
 #[cfg(test)]
 mod tests {
-    use crate::{FloatDiff, FloatEq};
-    use core::fmt;
+    use crate::FloatDiff;
 
     macro_rules! impl_tests {
         ($float:ident, $uint:ident, $nan_bits:ident) => {
@@ -215,223 +218,6 @@ mod tests {
 
                     //TODO: NaNs?
                 }
-
-                #[test]
-                fn eq_abs() {
-                    let eq = <$float as FloatEq>::eq_abs;
-                    let ne = <$float as FloatEq>::ne_abs;
-
-                    let check_eq = |a, b, max_diff| {
-                        check(eq, ne, a, b, max_diff, true);
-                        check(eq, ne, b, a, max_diff, true);
-
-                        assert!(float_eq!(a, b, abs <= max_diff));
-                        assert!(float_eq!(a, b, abs_all <= max_diff));
-
-                        assert!(!float_ne!(a, b, abs <= max_diff));
-                        assert!(!float_ne!(a, b, abs_all <= max_diff));
-
-                        assert_float_eq!(a, b, abs <= max_diff);
-                        assert_float_eq!(a, b, abs_all <= max_diff);
-
-                        // trailing comma
-                        assert!(float_eq!(a, b, abs <= max_diff,));
-                        assert_float_eq!(a, b, abs <= max_diff,);
-                    };
-
-                    let check_ne = |a, b, max_diff| {
-                        check(eq, ne, a, b, max_diff, false);
-                        check(eq, ne, b, a, max_diff, false);
-
-                        assert!(!float_eq!(a, b, abs <= max_diff));
-                        assert!(!float_eq!(a, b, abs_all <= max_diff));
-
-                        assert!(float_ne!(a, b, abs <= max_diff));
-                        assert!(float_ne!(a, b, abs_all <= max_diff));
-
-                        assert_float_ne!(a, b, abs <= max_diff);
-                        assert_float_ne!(a, b, abs_all <= max_diff);
-
-                        // trailing comma
-                        assert!(float_ne!(a, b, abs <= max_diff,));
-                        assert_float_ne!(a, b, abs <= max_diff,);
-                    };
-
-                    // useful in range where epsilon is relevent
-                    let one: $float = 1.;
-                    let eps = $float::EPSILON;
-                    check_eq(one, one, 0.);
-                    check_eq(one, one, eps);
-
-                    check_ne(one, one + eps, 0.);
-                    check_eq(one, one + eps, eps);
-                    check_ne(one, one + (2. * eps), eps);
-
-                    check_ne(one, one - eps, 0.);
-                    check_eq(one, one - eps, eps);
-                    check_ne(one, one - (2. * eps), eps);
-
-                    // unintuitive outside of range where epsilon is relevant
-                    let two: $float = 2.;
-                    let next = $float::from_bits(two.to_bits() + 1);
-                    check_ne(two, next, eps);
-                    check_eq(two, next, next - two);
-
-                    let a: $float = -128_000.;
-                    let next = $float::from_bits(a.to_bits() + 1);
-                    check_ne(a, next, eps);
-                    check_eq(a, next, a - next);
-
-                    // zero
-                    let zero = 0.;
-                    let neg_zero = -0.;
-                    check_eq(zero, neg_zero, 0.);
-
-                    // NaN
-                    let nans = nan_test_values();
-                    for &a in &nans {
-                        check_ne(one, a, eps);
-                        check_ne(a, one, eps);
-                        for &b in &nans {
-                            check_ne(a, b, eps);
-                        }
-                    }
-                }
-
-                #[test]
-                fn eq_rel() {
-                    let eq = <$float as FloatEq>::eq_rel;
-                    let ne = <$float as FloatEq>::ne_rel;
-
-                    let check_eq = |a, b, max_diff| {
-                        check(eq, ne, a, b, max_diff, true);
-                        check(eq, ne, b, a, max_diff, true);
-
-                        assert!(float_eq!(a, b, rel <= max_diff));
-                        assert!(float_eq!(a, b, rel_all <= max_diff));
-
-                        assert!(!float_ne!(a, b, rel <= max_diff));
-                        assert!(!float_ne!(a, b, rel_all <= max_diff));
-
-                        assert_float_eq!(a, b, rel <= max_diff);
-                        assert_float_eq!(a, b, rel_all <= max_diff);
-                    };
-
-                    let check_ne = |a, b, max_diff| {
-                        check(eq, ne, a, b, max_diff, false);
-                        check(eq, ne, b, a, max_diff, false);
-
-                        assert!(!float_eq!(a, b, rel <= max_diff));
-                        assert!(!float_eq!(a, b, rel_all <= max_diff));
-
-                        assert!(float_ne!(a, b, rel <= max_diff));
-                        assert!(float_ne!(a, b, rel_all <= max_diff));
-
-                        assert_float_ne!(a, b, rel <= max_diff);
-                        assert_float_ne!(a, b, rel_all <= max_diff);
-                    };
-
-                    // useful in range where epsilon is relevent
-                    let one: $float = 1.;
-                    let eps = $float::EPSILON;
-                    check_eq(one, one, 0.);
-                    check_eq(one, one, eps);
-
-                    check_ne(one, one + eps, 0.);
-                    check_eq(one, one + eps, eps);
-                    check_ne(one, one + (2. * eps), eps);
-
-                    check_ne(one, one - eps, 0.);
-                    check_eq(one, one - eps, eps);
-                    check_ne(one, one - (2. * eps), eps);
-
-                    // intuitive outside of range where epsilon is relevant
-                    let two: $float = 2.;
-                    let next = $float::from_bits(two.to_bits() + 1);
-                    check_eq(two, next, eps);
-
-                    let a: $float = -128_000.;
-                    let next = $float::from_bits(a.to_bits() + 1);
-                    check_eq(a, next, eps);
-
-                    // zero
-                    let zero = 0.;
-                    let neg_zero = -0.;
-                    check_eq(zero, neg_zero, 0.);
-
-                    // NaN
-                    let nans = nan_test_values();
-                    for &a in &nans {
-                        check_ne(one, a, eps);
-                        check_ne(a, one, eps);
-                        for &b in &nans {
-                            check_ne(a, b, eps);
-                        }
-                    }
-                }
-
-                #[test]
-                fn eq_ulps() {
-                    let eq = <$float as FloatEq>::eq_ulps;
-                    let ne = <$float as FloatEq>::ne_ulps;
-
-                    let check_eq = |a, b, max_diff| {
-                        check(eq, ne, a, b, max_diff, true);
-                        check(eq, ne, b, a, max_diff, true);
-                        assert!(float_eq!(a, b, ulps <= max_diff));
-                        assert!(!float_ne!(a, b, ulps <= max_diff));
-                        assert_float_eq!(a, b, ulps <= max_diff);
-                    };
-
-                    let check_ne = |a, b, max_diff| {
-                        check(eq, ne, a, b, max_diff, false);
-                        check(eq, ne, b, a, max_diff, false);
-                        assert!(!float_eq!(a, b, ulps <= max_diff));
-                        assert!(float_ne!(a, b, ulps <= max_diff));
-                        assert_float_ne!(a, b, ulps <= max_diff);
-                    };
-
-                    // useful in range where epsilon is relevent
-                    let one: $float = 1.;
-                    let eps = $float::EPSILON;
-                    check_eq(one, one, 0);
-                    check_eq(one, one, 1);
-
-                    check_ne(one, one + eps, 0);
-                    check_eq(one, one + eps, 1);
-                    check_ne(one, one + (2. * eps), 1);
-
-                    // unintuitive at the boundaries of powers of two
-                    check_ne(one, one - eps, 0);
-                    check_ne(one, one - eps, 1);
-                    check_eq(one, one - eps, 2);
-                    check_ne(one, one - (2. * eps), 1);
-
-                    // intuitive-ish outside of range where epsilon is relevant
-                    let two: $float = 2.;
-                    let next = $float::from_bits(two.to_bits() + 1);
-                    check_eq(two, next, 1);
-
-                    let a: $float = -128_000.;
-                    let next = $float::from_bits(a.to_bits() + 1);
-                    check_eq(a, next, 1);
-
-                    // zero
-                    let zero = 0.;
-                    let neg_zero = -0.;
-                    check_eq(zero, neg_zero, 0);
-                    check_eq(zero, neg_zero, 1);
-
-                    // NaN
-                    let nans = nan_test_values();
-                    for &a in &nans {
-                        check_ne(one, a, $uint::MAX);
-                        check_ne(a, one, $uint::MAX);
-                        for &b in &nans {
-                            check_ne(a, b, $uint::MAX);
-                        }
-                    }
-                }
             }
         };
     }
@@ -464,27 +250,4 @@ mod tests {
 
     impl_tests!(f32, u32, F32_NAN_BITS);
     impl_tests!(f64, u64, F64_NAN_BITS);
-
-    fn check<T, U, EQ, NE>(eq: EQ, ne: NE, a: T, b: T, max_diff: U, expect_equal: bool)
-    where
-        T: FloatEq + fmt::Display,
-        U: fmt::Display,
-        EQ: Fn(&T, &T, &U) -> bool,
-        NE: Fn(&T, &T, &U) -> bool,
-    {
-        assert!(
-            eq(&a, &b, &max_diff) == expect_equal,
-            "a: {}, b: {}, max_diff: {}",
-            a,
-            b,
-            max_diff
-        );
-        assert!(
-            ne(&a, &b, &max_diff) != expect_equal,
-            "a: {}, b: {}, max_diff: {}",
-            a,
-            b,
-            max_diff
-        );
-    }
 }